+++ ui/app/src/state/epics/system.ts
@@ -20,6 +20,7 @@ import {
   filter,
   ignoreElements,
   map,
+  mergeMap,
   startWith,
   switchMap,
   takeUntil,
@@ -52,7 +53,8 @@ import {
   showSystemNotification,
   showUnlockItemSuccessNotification,
   storeInitialized,
-  closeSiteSocket
+  closeSiteSocket,
+  emitSystemEvents
 } from '../actions/system';
 import { CrafterCMSEpic } from '../store';
 import {
@@ -81,6 +83,7 @@ import { batchActions, dispatchDOMEvent } from '../actions/misc';
 import StandardAction from '../../models/StandardAction';
 import { ProjectLifecycleEvent } from '../../models/ProjectLifecycleEvent';
 import { isDashboardAppUrl, isPreviewAppUrl, isProjectToolsAppUrl } from '../../utils/system';
+import { GlobalRoutes } from '../../env/routes';
 
 const msgs = defineMessages({
   siteSwitchedOnAnotherTab: {
@@ -140,11 +143,19 @@ const systemEpics: CrafterCMSEpic[] = [
   (action$) =>
     action$.pipe(
       ofType(emitSystemEvent.type),
+      tap(({ payload }) => getHostToHostBus().next(payload)),
+      map(({ payload: action }) => action)
+    ),
+  // endregion
+  // region emitSystemEvents
+  (action$: Observable<StandardAction<ReturnType<typeof emitSystemEvents>['payload']>>) =>
+    action$.pipe(
+      ofType(emitSystemEvents.type),
       tap(({ payload }) => {
         const hostToHost$ = getHostToHostBus();
-        hostToHost$.next(payload);
+        payload.events.forEach((action) => hostToHost$.next(action));
       }),
-      map(({ payload: action }) => action)
+      mergeMap((action) => action.payload.events)
     ),
   // endregion
   // region showDeleteItemSuccessNotification
@@ -475,7 +486,7 @@ const systemEpics: CrafterCMSEpic[] = [
             alert(
               formatMessage({ defaultMessage: "This project has been deleted, you'll be redirected to projects list." })
             );
-            window.location.href = `${authoringBase}#/sites`;
+            window.location.href = `${authoringBase}#${GlobalRoutes.Projects}`;
             return NEVER;
           } else {
             return [popSite({ siteId: site, isActive: true }), messageSharedWorker(closeSiteSocket({ site }))];